refactor: 로직을 커스텀 훅으로 분리하여 컴포넌트는 뷰 처리만 담당하도록 구조 개선 (#210)
